feat: change default path and remove dependency
This commit changes the default path for chronicle read and write files
from a cache directory to a user's home directory and removes the dependency on @cross/dir.

The previous default path was using a cache directory, which could be
cleared by the system or the user, resulting in data loss.
The new default path is `~/.chronicle/read` and `~/.chronicle/write`,
which is more persistent and user-friendly.

The dependency on `@cross/dir` has been removed because it is no longer
needed for determining the default path.
This reduces the number of dependencies and simplifies the codebase.